refactor: rename package from 'walelt' to 'wallet' for consistency
